+++ src/SASS/TodoList.scss
@@ -54,6 +54,6 @@
       }
 }
 
-.visible.menu.transition {
+.ui.dropdown .visible.menu.transition {
     position: absolute;
 }
